move standard error and test information functions due to circular import
