+++ adaptivetesting/math/estimators/__init__.py
@@ -4,3 +4,5 @@ from .__expect_a_posteriori import ExpectedAPosteriori
 from .__prior import Prior, NormalPrior, CustomPrior, CustomPriorException
 from .__functions.__estimators import probability_y0, probability_y1, maximize_likelihood_function, likelihood
 from .__functions.__bayes import maximize_posterior
+from .__standard_error import standard_error
+from .__test_information import test_information_function, item_information_function, prior_information_function
