+++ adaptivetesting/models/__adaptive_test.py
@@ -1,7 +1,7 @@
 from typing import List
 from .__test_item import TestItem
 from ..math.item_selection.__urrys_rule import urrys_rule
-from ..math.__standard_error import standard_error
+from ..math.estimators.__standard_error import standard_error
 import abc
 from .__test_result import TestResult
 from .__item_pool import ItemPool
